requirements.txt
Flask==2.3.2
yfinance==0.2.38
firebase-admin==6.3.0
python-dotenv==1.0.1
pandas==2.0.3 
matplotlib==3.7.3 